fix: Expands Telegram API endpoints and enables service integration
Allows Telegram handler activation when either legacy or gRPC service integration is enabled, resolving previous limitations.

Adds new Telegram API endpoints for checking binding status and unbinding via the frontend, improving user notification management.

Introduces a comprehensive integration test script for Telegram-related endpoints to ensure reliability.

Increases default gRPC timeout for Telegram service to accommodate slower responses.

Fixes type casting in binding status updates for improved database consistency.

+++ telegram-service/internal/database/store.go
@@ -149,8 +149,8 @@ func (s *Store) SetBindingStatus(ctx context.Context, userID string, status Bind
 		`UPDATE telegram_bindings
 		 SET status = $2,
 			blocked_reason = $3,
-			last_error_code = $4,
-			last_error_at = CASE WHEN $4 IS NULL THEN NULL ELSE NOW() END,
+			last_error_code = $4::integer,
+			last_error_at = CASE WHEN $4::integer IS NULL THEN NULL ELSE NOW() END,
 			updated_at = NOW()
 		 WHERE user_id = $1`,
 		userID, status, reason, code,
